requirements.txt


langchain
langchain-core
langchain-community
langchain-google-genai
langchain-pinecone
google-generativeai
pinecone
python-dotenv
